#334: add generic socker.io module + rm nodegit file in lib directory
