+++ server.js
@@ -4,13 +4,14 @@ config.rootFolder = __dirname;
 
 var app = require('./libs/express')(config);
 var logger = require('./libs/bunyan');
+var SocketIO = require('./libs/socket-io');
 
 require('./models/init')().then(function () {
   require('./routes')(app, config);
   var server = app.listen(config.port, function () {
     logger.info('Listening on port ' + config.port);
   });
-  var socket = require('./daemon/').init(server);
+  var socket = SocketIO.init(server);
 }).catch(function(err) {
   console.log(err);
 });
